fix: padronizacao dos IDs como long e padronizacao das saidas do AccessRepository, incluindo as alteracoes necessarias dos testes.

+++ app/src/main/java/com/cuscrud/data/local/dao/ProductDao.kt
@@ -18,7 +18,7 @@ interface ProdutoDao {
     fun getAllFlow(): Flow<List<ProdutoEntity>>
 
     @Query("SELECT * FROM produto WHERE id = :id")
-    suspend fun getById(id: Int): ProdutoEntity?
+    suspend fun getById(id: Long): ProdutoEntity?
 
     @Query("SELECT * FROM produto WHERE tipo = :tipoId")
     fun getByTipo(tipoId: Long): Flow<List<ProdutoEntity>>
